First commit with all necessary files
Some specific radicals can not be found within the Unihan database so their information is blank. Still working on this, but this version is functional for getting started with learning chinese character composition. Will go back and utilize another tool to define the radicals that do not currently have information.